Refactor IDs to Long and update repositories/services
Standardized entity IDs from Integer to Long across controllers, services, and repositories for MatierePremiere, Produit, and related entities. Added FournisseurPrix DTO and refactored DetailFournisseurRepository to use it. Introduced LogistiqueService for logistics operations. Updated presence and employee services to use new repository methods and types. Improved query methods and fixed type mismatches throughout the codebase.

+++ src/main/java/com/gestioncafe/controller/stock/MatierePremiereStockController.java
@@ -124,7 +124,7 @@ public class MatierePremiereStockController {
 
     @GetMapping("/edit/{id}")
     public String showEditForm(@PathVariable Long id, Model model) {
-        Optional<MatierePremiere> matierePremiere = matierePremiereRepository.findById(Math.toIntExact(id));
+        Optional<MatierePremiere> matierePremiere = matierePremiereRepository.findById((long) Math.toIntExact(id));
         if (matierePremiere.isPresent()) {
             model.addAttribute("matierePremiere", matierePremiere.get());
             model.addAttribute("unites", uniteRepository.findAll());
@@ -150,7 +150,7 @@ public class MatierePremiereStockController {
 
     @GetMapping("/delete/{id}")
     public String deleteMatierePremiere(@PathVariable Long id, RedirectAttributes redirectAttributes) {
-        matierePremiereRepository.deleteById(Math.toIntExact(id));
+        matierePremiereRepository.deleteById((long) Math.toIntExact(id));
         redirectAttributes.addFlashAttribute("success", "Matière première supprimée avec succès");
         return "redirect:/administratif/stock/matiere-premiere";
     }
